chore(backend): Rename assertValidSecretKey file to optionsAssertions
The optionsAssertions module will include assertion function for options
used across our package
